CircleSearchPC/requirements.txt
keyboard
mss
Pillow
pytesseract
google-generativeai
openai
translatepy